requirements.txt
flask
pandas
psycopg2
slack
slackclient
slackeventsapi
SQLAlchemy
GitPython
sandesh